+++ tests/fixtures/seeders/PostSeeder.php
@@ -1,8 +1,5 @@
 <?php
 
-namespace Database\Seeders;
-
-use App\Post;
 use Illuminate\Database\Seeder;
 
 class PostSeeder extends Seeder
@@ -14,6 +11,6 @@ class PostSeeder extends Seeder
      */
     public function run()
     {
-        Post::factory()->times(5)->create();
+        factory(\App\Post::class, 5)->create();
     }
 }
